+++ src/features/ImprovedScaleAndRotate.cpp
@@ -3,8 +3,77 @@
 #include <Geode/modify/GJRotationControl.hpp>
 #include <utils/Editor.hpp>
 #include <numbers>
+#include <span>
 
 using namespace geode::prelude;
+
+class ToggleSnapButton : public CCMenuItemSpriteExtra {
+protected:
+    ButtonSprite* m_sprite;
+    size_t m_ix = 0;
+    std::span<float> m_snaps;
+    std::function<void(float)> m_callback;
+
+    bool init(std::span<float> snaps, std::function<void(float)> callback) {
+        m_sprite = ButtonSprite::create("", "bigFont.fnt", "button-empty.png"_spr, .8f);
+        m_sprite->setScale(.65f);
+        if (!CCMenuItemSpriteExtra::init(m_sprite, nullptr, nullptr, nullptr))
+            return false;
+
+        m_snaps = snaps;
+        m_callback = callback;
+        this->updateSprite();
+        
+        return true;
+    }
+
+    void activate() override {
+        CCMenuItemSpriteExtra::activate();
+        m_ix += 1;
+        if (m_ix >= m_snaps.size()) {
+            m_ix = 0;
+        }
+        this->updateSprite();
+        if (m_callback) m_callback(m_snaps[m_ix]);
+    }
+
+    void updateSprite() {
+        m_sprite->setString(numToString(m_snaps[m_ix]).c_str());
+        m_sprite->m_label->setScale(.7f);
+		auto size = m_sprite->getScaledContentSize();
+		m_sprite->setPosition(size / 2);
+		m_sprite->setAnchorPoint(ccp(.5f, .5f));
+		this->setContentSize(size);
+    }
+
+public:
+    static ToggleSnapButton* create(std::span<float> snaps, std::function<void(float)> callback) {
+        auto ret = new ToggleSnapButton();
+        if (ret && ret->init(snaps, callback)) {
+            ret->autorelease();
+            return ret;
+        }
+        CC_SAFE_DELETE(ret);
+        return nullptr;
+    }
+
+    void setIndex(size_t index) {
+        m_ix = index;
+        this->updateSprite();
+        if (m_callback) m_callback(m_snaps[m_ix]);
+    }
+    float getValue() const {
+        return m_snaps[m_ix];
+    }
+
+    void setEnabled(bool enabled) override {
+        CCMenuItemSpriteExtra::setEnabled(enabled);
+        m_sprite->setCascadeColorEnabled(true);
+        m_sprite->setCascadeOpacityEnabled(true);
+        m_sprite->setColor(enabled ? ccWHITE : ccGRAY);
+        m_sprite->setOpacity(enabled ? 255 : 155);
+    }
+};
 
 // Returns radians
 static float angleOfPointOnCircle(CCPoint const& point) {
@@ -37,18 +106,6 @@ protected:
         this->setContentSize(ccp(210, 11));
         this->setAnchorPoint(ccp(.5f, .5f));
 
-        // Smaller slider option
-        // The conditionals here are horrendous but eh I was lazy
-        bool big = GameManager::get()->getGameVariable("0112");
-        for (size_t i = (big ? 2 : 3); i < (big ? 16 : 8); i += 1) {
-            auto x = m_obContentSize.width / (big ? 15 : 6) * (i - (big ? 1 : 2));
-            auto spr = CCSprite::createWithSpriteFrameName(
-                (i % 4) ? "slider-tick-small.png"_spr : "slider-tick.png"_spr
-            );
-            spr->setOpacity((i % 4) ? 105 : 255);
-            this->addChildAtPosition(spr, Anchor::Left, ccp(x, 0));
-        }
-
         return true;
     }
 
@@ -62,13 +119,48 @@ public:
         CC_SAFE_DELETE(ret);
         return nullptr;
     }
+
+    void updateSnapSize(size_t snap) {
+        this->removeAllChildren();
+
+        // Bigger slider (from 0.25-4.00) option
+        bool big = GameManager::get()->getGameVariable("0112");
+
+        // Big slider starts at 0.25, small slider starts at 0.50
+        // The + .06f is for a lil buffer to make sure there aren't lines way 
+        // too close to the edge
+        size_t start = ceilf(((big ? .25f : .50f) + .06f) * snap);
+
+        // Big slider ends at 4.00, small slider ends at 2.00
+        size_t count = (big ? snap * 4 : snap * 2);
+
+        float x = 42.f - 56.f / snap * (snap - start);
+        for (size_t i = start; i < count; i += 1) {
+            auto spr = CCSprite::createWithSpriteFrameName(
+                (i % snap) ? "slider-tick-small.png"_spr : "slider-tick.png"_spr
+            );
+            spr->setOpacity((i % snap) ? 105 : 255);
+            this->addChildAtPosition(spr, Anchor::Left, ccp(x, 0));
+            x += 56.f / snap;
+        }
+    }
 };
 
-class $modify(GJScaleControl) {
+class $modify(SnappableScaleControl, GJScaleControl) {
     $override
     bool init() {
         if (!GJScaleControl::init())
             return false;
+
+        auto snapLinesX = ScaleControlSnapLines::create();
+        m_sliderX->addChild(snapLinesX);
+        m_sliderX->m_touchLogic->setZOrder(1);
+        auto snapLinesY = ScaleControlSnapLines::create();
+        m_sliderY->addChild(snapLinesY);
+        m_sliderY->m_touchLogic->setZOrder(1);
+        auto snapLinesXY = ScaleControlSnapLines::create();
+        m_sliderXY->addChild(snapLinesXY);
+        m_sliderXY->m_touchLogic->setZOrder(1);
         
         m_scaleXLabel->setPositionX(-30);
         m_scaleYLabel->setPositionX(-30);
@@ -115,27 +207,35 @@ class $modify(GJScaleControl) {
 
         auto menu = m_scaleLockButton->getParent();
         menu->setID("lock-menu"_spr);
-        menu->setContentWidth(50);
+        menu->setContentWidth(75);
+        menu->setLayout(RowLayout::create());
 
         labelLockSpr(m_scaleLockButton->getNormalImage(), "Pos");
 
         auto snapBtn = CCMenuItemToggler::create(
             CCSprite::createWithSpriteFrameName("warpLockOffBtn_001.png"),
             CCSprite::createWithSpriteFrameName("warpLockOnBtn_001.png"),
-            this, nullptr
+            this, menu_selector(SnappableScaleControl::onSnapLock)
         );
+        snapBtn->toggle(false);
         labelLockToggle(snapBtn, "Snap");
         snapBtn->setID("snap-lock"_spr);
         menu->addChild(snapBtn);
 
-        menu->setLayout(RowLayout::create());
-
-        m_sliderX->addChild(ScaleControlSnapLines::create());
-        m_sliderX->m_touchLogic->setZOrder(1);
-        m_sliderY->addChild(ScaleControlSnapLines::create());
-        m_sliderY->m_touchLogic->setZOrder(1);
-        m_sliderXY->addChild(ScaleControlSnapLines::create());
-        m_sliderXY->m_touchLogic->setZOrder(1);
+        static std::array SNAPS { 1.f / 3, .25f, .2f, 1.f / 6, .1f };
+        auto snapSizeBtn = ToggleSnapButton::create(
+            SNAPS, [snapLinesX, snapLinesY, snapLinesXY](float snap) {
+                snapLinesX->updateSnapSize(snap);
+                snapLinesY->updateSnapSize(snap);
+                snapLinesXY->updateSnapSize(snap);
+            }
+        );
+        snapSizeBtn->setEnabled(false);
+        snapSizeBtn->setIndex(1);
+        snapSizeBtn->setID("snap-lock-size"_spr);
+        menu->addChild(snapSizeBtn);
+    
+        menu->updateLayout();
 
         return true;
     }
@@ -175,6 +275,10 @@ class $modify(GJScaleControl) {
         labelLockSpr(m_scaleLockButton->getNormalImage(), "Pos");
     }
 
+    void onSnapLock(CCObject* sender){
+        this->getSnapSizeBtn()->setEnabled(!static_cast<CCMenuItemToggler*>(sender)->m_toggled);
+    }
+
     $override
     void ccTouchMoved(CCTouch* touch, CCEvent* event) {
         // Call original without the delegate so it doesn't fire an unnecessary 
@@ -191,9 +295,11 @@ class $modify(GJScaleControl) {
         auto scaleX = this->scaleFromValue(m_sliderX->getThumb()->getValue());
         auto scaleY = this->scaleFromValue(m_sliderY->getThumb()->getValue());
         auto ratio = scaleY / scaleX;
-        if (auto lock = this->getSnapLock(); lock && lock->isToggled()) {
-            scaleX = roundf(scaleX / .25f) * .25f;
-            scaleY = roundf(scaleY / .25f) * .25f;
+        auto snapBtn = this->getSnapSizeBtn();
+        if (auto lock = this->getSnapLock(); lock && snapBtn && lock->isToggled()) {
+            auto snap = snapBtn->getValue();
+            scaleX = roundf(scaleX / snap) * snap;
+            scaleY = roundf(scaleY / snap) * snap;
         }
 
         if (m_scaleButtonType == 0) {
@@ -261,6 +367,11 @@ class $modify(GJScaleControl) {
             "hjfod.betteredit/lock-menu hjfod.betteredit/snap-lock"
         ));
     }
+    ToggleSnapButton* getSnapSizeBtn() {
+        return static_cast<ToggleSnapButton*>(this->querySelector(
+            "hjfod.betteredit/lock-menu hjfod.betteredit/snap-lock-size"
+        ));
+    }
 };
 
 class $modify(InputRotationControl, GJRotationControl) {
@@ -269,7 +380,6 @@ class $modify(InputRotationControl, GJRotationControl) {
         if (!GJRotationControl::init())
             return false;
         
-        // todo: you can place blocks through these
         auto input = TextInput::create(50, "Num");
         input->setScale(.8f);
         input->setID("input-angle"_spr);
@@ -286,7 +396,7 @@ class $modify(InputRotationControl, GJRotationControl) {
         this->addChild(input);
 
         auto menu = CCMenu::create();
-        menu->setContentWidth(50);
+        menu->setContentWidth(75);
         menu->setID("lock-menu"_spr);
 
         auto posBtn = CCMenuItemToggler::create(
@@ -301,11 +411,18 @@ class $modify(InputRotationControl, GJRotationControl) {
         auto snapBtn = CCMenuItemToggler::create(
             CCSprite::createWithSpriteFrameName("warpLockOffBtn_001.png"),
             CCSprite::createWithSpriteFrameName("warpLockOnBtn_001.png"),
-            this, nullptr
+            this, menu_selector(InputRotationControl::onSnapLock)
         );
         labelLockToggle(snapBtn, "Snap");
         snapBtn->setID("snap-lock"_spr);
         menu->addChild(snapBtn);
+
+        static std::array SNAPS { 10.f, 15.f, 18.f };
+        auto snapSizeBtn = ToggleSnapButton::create(SNAPS, nullptr);
+        snapSizeBtn->setEnabled(false);
+        snapSizeBtn->setIndex(1);
+        snapSizeBtn->setID("snap-lock-size"_spr);
+        menu->addChild(snapSizeBtn);
 
         menu->setLayout(RowLayout::create());
         menu->setPosition(110, 35);
@@ -318,11 +435,12 @@ class $modify(InputRotationControl, GJRotationControl) {
     void draw() {
         GJRotationControl::draw();
         // Draw large ticks every 45° and small ticks every 15°
-        for (size_t i = 0; i < 24; i += 1) {
-            glLineWidth((i % 3) ? 1 : 2);
-            float len = (i % 3) ? 2.5f : 5;
-            float angle = i * 15;
-            ccDrawLine(pointOnCircle(angle, 60 - len), pointOnCircle(angle, 60 + len));
+        auto tickSize = this->getSnapSizeBtn()->getValue();
+        for (float angle = 0; angle < 360; angle += tickSize) {
+            bool isBigTick = size_t(angle) % 45;
+            glLineWidth(isBigTick ? 1 : 2);
+            float len = isBigTick ? 2.5f : 5;
+            ccDrawLine(pointOnCircle(angle + 90, 60 - len), pointOnCircle(angle + 90, 60 + len));
         }
     }
 
@@ -340,7 +458,8 @@ class $modify(InputRotationControl, GJRotationControl) {
 
         auto angle = this->getThumbValue();
         if (auto lock = this->getSnapLock(); lock && lock->isToggled()) {
-            angle = roundf(this->getThumbValue() / 15) * 15;
+            auto tickSize = this->getSnapSizeBtn()->getValue();
+            angle = roundf(this->getThumbValue() / tickSize) * tickSize;
             m_controlSprite->setPosition(pointOnCircle(-angle + 90, 60));
         }
         m_delegate->angleChanged(angle);
@@ -354,6 +473,10 @@ class $modify(InputRotationControl, GJRotationControl) {
         this->getInput()->setString(numToString(angle, 3));
 
         be::evilForceTouchPrio(static_cast<EditorUI*>(m_delegate), this->getInput()->getInputNode());
+    }
+
+    void onSnapLock(CCObject* sender){
+        this->getSnapSizeBtn()->setEnabled(!static_cast<CCMenuItemToggler*>(sender)->m_toggled);
     }
 
     float getThumbValue() const {
@@ -376,6 +499,12 @@ class $modify(InputRotationControl, GJRotationControl) {
     CCMenuItemToggler* getPosLock() {
         return static_cast<CCMenuItemToggler*>(this->querySelector(
             "hjfod.betteredit/lock-menu hjfod.betteredit/pos-lock"
+        ));
+    }
+    // todo: this should probably be a field instead for efficiency
+    ToggleSnapButton* getSnapSizeBtn() {
+        return static_cast<ToggleSnapButton*>(this->querySelector(
+            "hjfod.betteredit/lock-menu hjfod.betteredit/snap-lock-size"
         ));
     }
 };